[Refactor] #9 상세 조회 기능에 대해 인터페이스 / 구현체를 분리하여 확장성 및 유연성 확보
- 각 기능별로 인터페이스 / 구현체를 분리하여 사용함으로써 다형성을 최대한 활용
- 네이밍 -> 인터페이스 : 도메인+기능+Service / 구현체 -> 도메인+기능+ServiceImpl
