+++ src/main/java/com/kuit/findyou/domain/report/service/facade/ReportServiceFacade.java
@@ -1,0 +1,9 @@
+package com.kuit.findyou.domain.report.service.facade;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ReportService {
+}
